Move remaining functions from detect to detect_utils
Remove test_detect
