+++ hardware/detect.py
@@ -18,8 +18,6 @@ import argparse
 import json
 import os
 import pprint
-from subprocess import PIPE
-from subprocess import Popen
 import sys
 
 from hardware import areca
@@ -36,60 +34,6 @@ from hardware import megacli
 from hardware import rtc
 from hardware import sensors
 from hardware import system
-
-
-AUXV_FLAGS = ["AT_HWCAP", "AT_HWCAP2", "AT_PAGESZ",
-              "AT_FLAGS", "AT_PLATFORM"]
-# These flags may or not be present on a particular arch
-AUXV_OPT_FLAGS = ["AT_BASE_PLATFORM"]
-
-
-def detect_auxv(hw_lst):
-    new_env = os.environ.copy()
-    new_env["LD_SHOW_AUXV"] = "1"
-
-    cmd = Popen("/bin/true",
-                env=new_env,
-                stdout=PIPE)
-    stdout, err = cmd.communicate()
-    if err is not None:
-        sys.stderr.write("Info: AUXV output received\n")
-        return
-
-    auxv = dict()
-    supported_flags = AUXV_FLAGS + AUXV_OPT_FLAGS
-    for line in stdout.decode("utf-8").splitlines():
-        k, v = [i.strip() for i in line.split(":")]
-        if k in supported_flags:
-            auxv[k[3:].lower()] = v
-            hw_lst.append(('hw', 'auxv', k[3:].lower(), v))
-
-
-def parse_ahci(hrdw, words):
-    if len(words) < 4:
-        return
-    if "flags" in words[2]:
-        flags = ""
-        for flag in sorted(words[3:]):
-            flags = "%s %s" % (flags, flag)
-        hrdw.append(('ahci', words[1], "flags", flags.strip()))
-
-
-def parse_dmesg(hrdw):
-    """Run dmesg and parse the output."""
-
-    _, output = detect_utils.cmd("dmesg")
-    for line in output.split('\n'):
-        words = line.strip().split(" ")
-
-        if words[0].startswith("[") and words[0].endswith("]"):
-            words = words[1:]
-
-        if not words:
-            continue
-
-        if "ahci" in words[0]:
-            parse_ahci(hrdw, words)
 
 
 def parse_args(arguments):
@@ -140,8 +84,8 @@ def main():
     hrdw.extend(ipmi.get_ipmi_sdr())
     hrdw.extend(rtc.detect_rtc_clock())
 
-    detect_auxv(hrdw)
-    parse_dmesg(hrdw)
+    detect_utils.detect_auxv(hrdw)
+    detect_utils.parse_dmesg(hrdw)
     bios_hp.dump_hp_bios(hrdw)
 
     if args.benchmark:
